+++ src/game/constants.ts
@@ -3,6 +3,6 @@ import Phaser from "phaser";
 export const gameWidth = 430;
 export const gameHeight = 844;
 export const tankBounds = new Phaser.Geom.Rectangle(0, 0, 430, 844);
-export const controlPanelTop = 524;
+export const controlPanelTop = 62;
 export const toastX = gameWidth / 2;
 export const toastY = 100;